Cleaning up the local server dev mode a bit

+++ postBuild.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
-echo "Copying images and media to static public folder ..."
-cp ../Media/Instruction\ Images/*.png ./public/media/
+
+# copy latest media images if found
+if [ -e ../Media/Instruction\ Images/ ]; then
+  echo "Copying images to static public folder ..."
+  cp ../Media/Instruction\ Images/*.png ./public/media/
+fi
+
 if [ -e ../ExtensionRefactor/KarunaEEC-Extension.zip ]; then
+  echo "Copying extension zip to static public folder ..."
   cp ../ExtensionRefactor/KarunaEEC-Extension.zip ./public/media/
 fi
